requirements.txt
# ----------- CORE -----------
numpy==1.26.4
scipy==1.11.4
pandas==2.1.4
tqdm==4.66.1
pydantic==2.5.3
orjson==3.9.10
ujson==5.9.0
loguru==0.7.2
rich==13.7.0
python-dotenv>=1.0

# ----------- TORCH CPU -----------
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ----------- NLP -----------
transformers==4.37.2
tokenizers==0.15.2
sentence-transformers==2.5.1
accelerate==0.26.1
safetensors==0.4.2

# ----------- VECTOR SEARCH -----------
faiss-cpu==1.7.4
rank-bm25==0.2.2

# ----------- DOCUMENT PARSING -----------
pymupdf==1.23.7
python-docx==1.1.0
unstructured==0.11.8

# ----------- API -----------
fastapi==0.109.0
uvicorn==0.27.0.post1

# ----------- LLM -----------
openai>=1.0.0
httpx>=0.25.0

# ----------- RAG Benchmark (evaluation) -----------
rouge-score>=0.1.2
nltk>=3.8